Unit_Skill
Add : Buff, Util,Debuff
Update : GameManager
버프 : 공격력, 공격속도, 이동속도,다른라인공격력,
디버프 : 공격력, 공격속도, 라인바꾸기
유틸 : 저코스트 유닛 잡아먹고 쌔지기
, 저코스트 유닛 소환

테스트는 하나도 안됌

+++ Assets/CSV/CSVtoSO.cs
@@ -18,7 +18,7 @@ public class CSVtoSO
             Unit unit = ScriptableObject.CreateInstance<Unit>();
             unit.unitName = splitData[0];
             unit.cost = int.Parse(splitData[1]);
-            unit.damage = float.Parse(splitData[2]);
+            unit.damage = int.Parse(splitData[2]);
             unit.attackSpeed = float.Parse(splitData[3]);
             unit.speed = float.Parse(splitData[4]);
             unit.imagePath = splitData[5];
